Implement dev review feedback on the reactive app
