--- finish/bff/src/main/java/io/openliberty/guides/bff/BffApplication.java
@@ -1,22 +1,0 @@
-// tag::copyright[]
-/*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - Initial implementation
- *******************************************************************************/
- // end::copyright[]
-package io.openliberty.guides.bff;
-
-// JAX-RS
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("api")
-public class BffApplication extends Application {
-
-}
