--- finish/bff/src/main/java/io/openliberty/guides/bff/model/JobModel.java
@@ -1,13 +1,0 @@
-package io.openliberty.guides.bff.model;
-
-public class JobModel {
-    private String jobId;
-
-    public String getJobId() {
-        return jobId;
-    }
-
-    public void setJobId(String jobId) {
-        this.jobId = jobId;
-    }
-}
