--- finish/bff/src/main/java/io/openliberty/guides/bff/model/JobsModel.java
@@ -1,16 +1,0 @@
-package io.openliberty.guides.bff.model;
-
-import java.util.List;
-
-public class JobsModel {
-    private List<JobResultModel> results;
-
-    public List<JobResultModel> getResults() {
-        return results;
-    }
-
-    public void setResults(List<JobResultModel> results) {
-        this.results = results;
-    }
-    
-}
