--- finish/bff/src/main/liberty/config/server.xml
@@ -1,16 +1,0 @@
-<server description="Sample Liberty server">
-
-  <featureManager>
-    <feature>jaxrs-2.1</feature>
-    <feature>cdi-2.0</feature>
-    <feature>mpRestClient-1.2</feature>
-    <feature>mpConfig-1.3</feature>
-    <feature>jsonp-1.1</feature>
-    <feature>concurrent-1.0</feature>
-    <feature>jsonb-1.0</feature>
-  </featureManager>
-
-  <httpEndpoint host="*" httpPort="${default.http.port}" httpsPort="${default.https.port}" id="defaultHttpEndpoint"/>
-
-  <webApplication location="bff.war" contextRoot="/"/>
-</server>
