+++ finish/gateway/pom.xml
@@ -11,7 +11,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>inventory</artifactId>
+    <artifactId>gateway</artifactId>
     <packaging>war</packaging>
 
     <dependencies>
@@ -57,8 +57,8 @@
             <artifactId>junit</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>testcontainers</artifactId>
+            <groupId>org.mock-server</groupId>
+            <artifactId>mockserver-netty</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.cxf</groupId>
@@ -152,6 +152,19 @@
                         <id>package-server</id>
                         <configuration>
                             <packageFile>${project.build.directory}/defaultServer.zip</packageFile>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>test-start-server</id>
+                        <phase>pre-integration-test</phase>
+                        <goals>
+                            <goal>test-start-server</goal>
+                        </goals>
+                        <configuration>
+                            <jvmOptions>
+                                <param>-Dio.openliberty.guides.gateway.client.JobClient/mp-rest/url=http://localhost:9082</param>
+                                <param>-Dio.openliberty.guides.gateway.client.InventoryClient/mp-rest/url=http://localhost:9082</param>
+                            </jvmOptions>
                         </configuration>
                     </execution>
                 </executions>
