+++ finish/gateway/src/main/java/io/openliberty/guides/gateway/GatewayApplication.java
@@ -1,0 +1,22 @@
+// tag::copyright[]
+/*******************************************************************************
+ * Copyright (c) 2017 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - Initial implementation
+ *******************************************************************************/
+ // end::copyright[]
+package io.openliberty.guides.gateway;
+
+// JAX-RS
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("api")
+public class GatewayApplication extends Application {
+
+}
