+++ finish/job/pom.xml
@@ -15,6 +15,11 @@
     <packaging>war</packaging>
 
     <dependencies>
+        <dependency>
+            <groupId>io.openliberty.guides</groupId>
+            <artifactId>models</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
        <dependency>
            <groupId>io.openliberty.features</groupId>
            <artifactId>concurrent-1.0</artifactId>
