+++ finish/models/pom.xml
@@ -1,0 +1,15 @@
+<?xml version='1.0' encoding='utf-8'?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>io.openliberty.guides</groupId>
+  <artifactId>models</artifactId>
+  <version>1.0-SNAPSHOT</version>
+
+  <parent>
+    <groupId>io.openliberty.guides</groupId>
+    <artifactId>kube-demo</artifactId>
+    <version>1.0-SNAPSHOT</version>
+  </parent>
+
+</project>
