+++ finish/models/src/main/java/io/openliberty/guides/models/JobModel.java
@@ -1,0 +1,13 @@
+package io.openliberty.guides.models;
+
+public class JobModel {
+    private String jobId;
+
+    public String getJobId() {
+        return jobId;
+    }
+
+    public void setJobId(String jobId) {
+        this.jobId = jobId;
+    }
+}
