+++ finish/models/src/main/java/io/openliberty/guides/models/JobsModel.java
@@ -1,0 +1,16 @@
+package io.openliberty.guides.models;
+
+import java.util.List;
+
+public class JobsModel {
+    private List<JobResultModel> results;
+
+    public List<JobResultModel> getResults() {
+        return results;
+    }
+
+    public void setResults(List<JobResultModel> results) {
+        this.results = results;
+    }
+    
+}
