+++ finish/scripts/stop-app
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-docker stop bff
+docker stop gateway
 docker stop job-service
 docker stop inventory-service
 docker stop system-service
