+++ finish/system/pom.xml
@@ -16,6 +16,11 @@
 
     <dependencies>
        <dependency>
+           <groupId>io.openliberty.guides</groupId>
+           <artifactId>models</artifactId>
+           <version>1.0-SNAPSHOT</version>
+       </dependency>
+       <dependency>
            <groupId>io.openliberty.features</groupId>
            <artifactId>concurrent-1.0</artifactId>
            <type>esa</type>
